factor out layers into individual files
Closes issue #4
