+++ layers/static.js
@@ -1,0 +1,54 @@
+var path = require('path');
+var parse = require('parseurl');
+var utilsPause = require('pause');
+var send = require('send');
+/**
+ * Static
+ *
+ * Serves up static page (if it exists).
+ *
+ */
+module.exports = function static (req, res, next) {
+  var options  = {}
+  var redirect = true
+
+  if ('GET' != req.method && 'HEAD' != req.method) return next();
+  var pathn = parse(req).pathname;
+  var pause = utilsPause(req);
+
+  function resume() {
+    next();
+    pause.resume();
+  }
+
+  function directory() {
+
+    if (!redirect) return resume();
+    var pathname = url.parse(req.originalUrl).pathname;
+    res.statusCode = 301;
+    res.setHeader('Location', pathname + '/');
+    res.end('Redirecting to ' + utilsEscape(pathname) + '/');
+  }
+
+  function error(err) {
+    if (404 == err.status){
+      // look for implicit `*.html` if we get a 404
+      return path.extname(err.path) === ''
+        ? serve(pathn + ".html")
+        : resume()
+    }
+    next(err);
+  }
+
+  var serve = function(pathn){
+    send(req, pathn, {
+        maxage: options.maxAge || 0,
+        root: req.setup.publicPath,
+        hidden: options.hidden
+      })
+      .on('error', error)
+      .on('directory', directory)
+      .pipe(res)
+  }
+  serve(pathn)
+}
