+++ layers/underscore.js
@@ -1,0 +1,15 @@
+var shouldIgnore = require('../lib/shouldignore');
+var notFound = require('./fallbacks').notFound;
+/**
+ * Underscore
+ *
+ * Returns 404 if path contains beginning underscore
+ *
+ */
+module.exports = function underscore (req, rsp, next){
+  if(shouldIgnore(req.url)){
+    notFound(req, rsp, next)
+  }else{
+    next()
+  }
+}
